fixing icon location
