+++ src/App.js
@@ -2,7 +2,7 @@ import { Route, Routes } from "react-router-dom";
 import "./App.css";
 import AppLayout from "./layout/AppLayout";
 import MainMapPage from "./pages/MainMap/MainMapPage";
-import SmallMapPage from "./pages/SmallMap/SmallMapPage";
+import AttractionMapPage from "./pages/AttractionMap/AttractionMapPage";
 import InfoPage from "./pages/Info/InfoPage";
 import InfoDetailPage from "./pages/InfoDetail/InfoDetailPage";
 import HouseMapPage from "./pages/HouseMap/HouseMapPage";
@@ -13,7 +13,7 @@ function App() {
       <Routes>
         <Route path="/" element={<AppLayout/>}> {/*네브바*/}
           <Route index element={<MainMapPage/>}/> {/*전체 지도*/}
-          <Route path="attraction/:id" element={<SmallMapPage/>}/> {/*관광지 지도*/}
+          <Route path="attraction/:id" element={<AttractionMapPage/>}/> {/*관광지 지도*/}
           <Route path="house/:id" element={<HouseMapPage/>}/>
           <Route path="info">
             <Route index element={<InfoPage/>}/> {/*전체 정보 페이지*/}
